feat: retrieve and show saved records - add card view to show saved records - use third-party library to show circular thumbnail - fetch saved places and show them in recycle adapter - add activity to show full details

+++ HappyPlaces/app/src/main/java/com/shail_singh/happyplaces/database/DatabaseHandler.kt
@@ -23,7 +23,7 @@ class DatabaseHandler(application: Application) {
         return happyPlaceDao.fetchAll()
     }
 
-    fun fetchPlace(id: Int): Flow<HappyPlaceModel> {
+    fun fetchPlace(id: Long): Flow<HappyPlaceModel> {
         return happyPlaceDao.fetchRecord(id)
     }
